# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorDetail:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      message:
        example: Invalid input data
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorDetail'
    type: object
...
  dto.MoodleConnectionResponse:
    properties:
      connected:
        type: boolean
      error:
        type: string
      message:
        type: string
    type: object
  dto.MoodleCourse:
    properties:
      id:
        type: string
      name:
        type: string
      short_name:
        type: string
    type: object
  dto.MoodleCoursesResponse:
    properties:
      courses:
        items:
          $ref: '#/definitions/dto.MoodleCourse'
        type: array
    type: object
  dto.MoodleSyncResponse:
    properties:
      course_id:
        type: string
      message:
        type: string
      moodle_id:
        type: string
    type: object
...
  dto.SyncMoodleRequest:
    properties:
      course_name:
        type: string
    required:
    - course_name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@testgen.com
    name: API Support
  description: API for test generation system with LLM integration and Moodle export
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Test Generation System API
  version: "1.0"
paths:
...
  /moodle/courses:
    get:
      description: Retrieve list of available courses from Moodle
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MoodleCoursesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to retrieve courses
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Moodle courses
      tags:
      - moodle
  /moodle/validate:
    get:
      description: Check if the Moodle server connection is working properly
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MoodleConnectionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "503":
          description: Connection failed
          schema:
            $ref: '#/definitions/dto.MoodleConnectionResponse'
      security:
      - BearerAuth: []
      summary: Validate Moodle connection
      tags:
      - moodle
...
  /tests/{id}/export-xml:
    get:
      description: Export a test and its questions to Moodle XML format for download
      parameters:
      - description: Test ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/xml
      responses:
        "200":
          description: XML file
          schema:
            type: string
        "400":
          description: Invalid test ID or test has no questions
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Test not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Export failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export test to Moodle XML format
      tags:
      - moodle
  /tests/{id}/sync-moodle:
    post:
      consumes:
      - application/json
      description: Synchronize a test with Moodle by uploading it as a quiz to a specified
        course
      parameters:
      - description: Test ID
        in: path
        name: id
        required: true
        type: string
      - description: Sync request with course name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SyncMoodleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MoodleSyncResponse'
        "400":
          description: Invalid input or test has no questions
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Test not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Sync failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Sync test to Moodle
      tags:
      - moodle
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"